Refactor subscripting to standalone source files
