+++ GenericJSON/QueryingTests.swift
@@ -37,48 +37,4 @@ class QueryingTests: XCTestCase {
         XCTAssertEqual(JSON.null.isNull, true)
         XCTAssertEqual(JSON.string("foo").isNull, false)
     }
-
-    func testArraySubscripting() {
-        let json: JSON = ["foo", "bar"]
-        XCTAssertEqual(json[0], JSON.string("foo"))
-        XCTAssertEqual(json[-1], nil)
-        XCTAssertEqual(json[2], nil)
-        XCTAssertEqual(json["foo"], nil)
-    }
-
-    func testStringSubscripting() {
-        let json: JSON = ["foo": "bar"]
-        XCTAssertEqual(json["foo"], JSON.string("bar"))
-        XCTAssertEqual(json[0], nil)
-        XCTAssertEqual(json["nonesuch"], nil)
-    }
-
-    func testStringSubscriptingSugar() {
-        let json: JSON = ["foo": "bar"]
-        XCTAssertEqual(json.foo, JSON.string("bar"))
-        XCTAssertEqual(json.nonesuch, nil)
-    }
-    
-    func testKeyPath() {
-        let json: JSON = [
-            "string": "foo bar",
-            "boolean": true,
-            "number": 123,
-            "object": [
-                "str": "col",
-                "arr": [1, 2, 3],
-                "obj": [
-                    "x": "rah",
-                    "y": "tar",
-                    "z": "yaz"
-                ]
-            ]
-        ]
-        XCTAssertEqual(json[keyPath: "string"], "foo bar")
-        XCTAssertEqual(json[keyPath: "boolean"], true)
-        XCTAssertEqual(json[keyPath: "number"], 123)
-        XCTAssertEqual(json[keyPath: "object.str"], "col")
-        XCTAssertEqual(json[keyPath: "object.arr"], [1, 2, 3])
-        XCTAssertEqual(json[keyPath: "object.obj.y"], "tar")
-    }
 }
